thank you page added

+++ main.py
@@ -325,46 +325,57 @@ def add_to_cart(product_id):
     cursor.close()
     connection.close()
 
+    return redirect("/cart")
 
+@app.route("/cart/<product_id>/remove", methods=["POST"])
+@login_required
+def remove(product_id):
 
+    connection = connect_db()
+    cursor = connection.cursor()
 
-@app.route("/checkout")
-def checkout():
-    return render_template("checkout.html.jinja")
+    cursor.execute("""
+        DELETE FROM Cart 
+        WHERE ProductID = %s AND UserID = %s
+    """, (product_id, current_user.id))
+
+    connection.commit() 
+    cursor.close()
+    connection.close()
+
+    return redirect('/cart')
+
 
 @app.route("/checkout", methods=["GET", "POST"])
 @login_required
 def checkout():
     connection = connect_db()
+
     cursor = connection.cursor()
     cursor.execute("""
-        SELECT * FROM `Cart`
-        JOIN `Product` ON `Product`.`ID` = `Cart`.`ProductID`
-        WHERE `UserID` = %s
-    """, (current_user.id,))
-    
+        SELECT * FROM `Cart` 
+        JOIN `Product` ON `Cart`.`ProductID` = `Product`.`ID`
+        WHERE `UserID` =  %s 
+        """, (current_user.id,) )
     result = cursor.fetchall()
-    if request.method == "POST":
-        cursor.execute(
-            "INSERT INTO `Sales` (`UserID`) VALUES (%s)",
-            (current_user.id,)
-        )
-        sales = cursor.lastrowid
-        for item in result:
-            cursor.execute(
-                "INSERT INTO `SalesCart` (`SalesID`, `ProductID`, `Quantity`) VALUES (%s, %s, %s)",
-                (sales, item['ProductID'], item['Quantity'])
-            )
 
-        cursor.execute(
-            "DELETE FROM `Cart` WHERE `UserID` = %s",
-            (current_user.id,)
-        )
-        connection.commit()
+    sale = cursor.lastrowid
+    if request.method == "POST":
+        # create the sale in the database
+        cursor.execute("INSERT INTO `Sale` (`UserID`) VALUES (%s)", (current_user.id,))
+        sale = cursor.lastrowid  # Retrieve the last inserted sale ID
+        # store products bought
+        for item in result:
+            cursor.execute("INSERT INTO `SaleProduct` (`SaleID`, `ProductID`, `Quantity`) VALUES (%s, %s, %s)", (sale, item['ID'], item['quantity']))
+        # empty cart
+        cursor.execute("DELETE FROM `Cart` WHERE `UserID` = %s", (current_user.id,))
+        # thank you screen
         return redirect('/thank_you')
+
     connection.close()
-    return render_template("checkout.html.jinja", cart=result)
- 
+
+    return render_template("checkout.html.jinja" , cart=result)
+
 
 @app.route("/thank_you")
 @login_required
